+++ Modules/Intent.Modules.ApplicationTemplate.Builder/Api/MinimumDependencyVersionsModel.cs
@@ -1,0 +1,83 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Intent.Metadata.Models;
+using Intent.Modules.Common;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.Api.ApiElementModel", Version = "1.0")]
+
+namespace Intent.Modules.ApplicationTemplate.Builder.Api
+{
+    [IntentManaged(Mode.Fully, Signature = Mode.Fully)]
+    public class MinimumDependencyVersionsModel : IMetadataModel, IHasStereotypes, IHasName, IElementWrapper
+    {
+        public const string SpecializationType = "Minimum Dependency Versions";
+        public const string SpecializationTypeId = "63605213-ab56-4042-b801-907bb5a79c08";
+        protected readonly IElement _element;
+
+        [IntentManaged(Mode.Fully)]
+        public MinimumDependencyVersionsModel(IElement element, string requiredType = SpecializationType)
+        {
+            if (!requiredType.Equals(element.SpecializationType, StringComparison.InvariantCultureIgnoreCase))
+            {
+                throw new Exception($"Cannot create a '{GetType().Name}' from element with specialization type '{element.SpecializationType}'. Must be of type '{SpecializationType}'");
+            }
+            _element = element;
+        }
+
+        public string Id => _element.Id;
+
+        public string Name => _element.Name;
+
+        public string Comment => _element.Comment;
+
+        public IEnumerable<IStereotype> Stereotypes => _element.Stereotypes;
+
+        public IElement InternalElement => _element;
+
+        public IList<ModuleModel> Modules => _element.ChildElements
+            .GetElementsOfType(ModuleModel.SpecializationTypeId)
+            .Select(x => new ModuleModel(x))
+            .ToList();
+
+        public override string ToString()
+        {
+            return _element.ToString();
+        }
+
+        public bool Equals(MinimumDependencyVersionsModel other)
+        {
+            return Equals(_element, other?._element);
+        }
+
+        public override bool Equals(object obj)
+        {
+            if (ReferenceEquals(null, obj)) return false;
+            if (ReferenceEquals(this, obj)) return true;
+            if (obj.GetType() != this.GetType()) return false;
+            return Equals((MinimumDependencyVersionsModel)obj);
+        }
+
+        public override int GetHashCode()
+        {
+            return (_element != null ? _element.GetHashCode() : 0);
+        }
+    }
+
+    [IntentManaged(Mode.Fully)]
+    public static class MinimumDependencyVersionsModelExtensions
+    {
+
+        public static bool IsMinimumDependencyVersionsModel(this ICanBeReferencedType type)
+        {
+            return type != null && type is IElement element && element.SpecializationTypeId == MinimumDependencyVersionsModel.SpecializationTypeId;
+        }
+
+        public static MinimumDependencyVersionsModel AsMinimumDependencyVersionsModel(this ICanBeReferencedType type)
+        {
+            return type.IsMinimumDependencyVersionsModel() ? new MinimumDependencyVersionsModel((IElement)type) : null;
+        }
+    }
+}