+++ Modules/Intent.Modules.Common.CSharp/Builder/CSharpClassMethod.cs
@@ -130,7 +130,7 @@ public class CSharpClassMethod : CSharpMember<CSharpClassMethod>, ICSharpMethodD
 
 
     /// <summary>
-    /// Use <see cref="AddOptionalCancellationTokenParameter"/> instead.
+    /// Use <see cref="AddOptionalCancellationTokenParameter()"/> instead.
     /// </summary>
     [Obsolete]
     public CSharpClassMethod AddOptionalCancellationTokenParameter<T>(CSharpTemplateBase<T> template) =>
@@ -142,6 +142,12 @@ public class CSharpClassMethod : CSharpMember<CSharpClassMethod>, ICSharpMethodD
         AddParameter(
             type: File.Template.UseType("System.Threading.CancellationToken"),
             name: "cancellationToken",
+            configure: parameter => parameter.WithDefaultValue("default"));
+
+    public CSharpClassMethod AddOptionalCancellationTokenParameter(string parameterName) =>
+        AddParameter(
+            type: File.Template.UseType("System.Threading.CancellationToken"),
+            name: parameterName,
             configure: parameter => parameter.WithDefaultValue("default"));
 
     public CSharpClassMethod AddGenericParameter(string typeName)